mimic-platform/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
langchain-groq
langchain
langchain-community
pinecone-client
requests
numpy
astropy
opencv-python
pillow
matplotlib
pywavelets
scikit-image
hypothesis
pytest
alembic